% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.warmstart.R
\name{bcf.warmstart}
\alias{bcf.warmstart}
\title{Fit warm-start Bayesian Causal Forests}
\usage{
bcf.warmstart(
  y,
  z,
  x_control,
  x_moderate = x_control,
  pihat,
  w = NULL,
  nburn,
  nsim,
  nthin = 1,
  update_interval = 100,
  ntree_control = 200,
  sd_control = 2 * sd(y),
  base_control = 0.95,
  power_control = 2,
  ntree_moderate = 50,
  sd_moderate = sd(y),
  base_moderate = 0.25,
  power_moderate = 3,
  nu = 3,
  lambda = NULL,
  sigq = 0.9,
  sighat = NULL,
  randeff = FALSE,
  include_pi = "control",
  use_muscale = TRUE,
  use_tauscale = TRUE,
  ini_bcf = FALSE,
  update_mu_loading_tree = FALSE,
  verbose = FALSE,
  n_cores = NULL,
  warm_start_fit = NULL,
  cutpoint_grid = "uniform"
)
}
\arguments{
\item{y}{Response variable}

\item{z}{Treatment variable}

\item{x_control}{Design matrix for the "prognostic" function mu(x)}

\item{x_moderate}{Design matrix for the covariate-dependent treatment effects tau(x)}

\item{pihat}{Length n estimates of}

\item{w}{An optional vector of weights. When present, BCF fits a model \eqn{y | x ~ N(f(x), \sigma^2 / w)}, where \eqn{f(x)} is the unknown function.}

\item{nburn}{Number of burn-in MCMC iterations}

\item{nsim}{Number of MCMC iterations to save after burn-in}

\item{nthin}{Save every nthin'th MCMC iterate. The total number of MCMC iterations will be nsim*nthin + nburn.}

\item{update_interval}{Print status every update_interval MCMC iterations}

\item{ntree_control}{Number of trees in mu(x)}

\item{sd_control}{SD(mu(x)) marginally at any covariate value (or its prior median if use_muscale=TRUE)}

\item{base_control}{Base for tree prior on mu(x) trees (see details)}

\item{power_control}{Power for the tree prior on mu(x) trees}

\item{ntree_moderate}{Number of trees in tau(x)}

\item{sd_moderate}{SD(tau(x)) marginally at any covariate value (or its prior median if use_tauscale=TRUE)}

\item{base_moderate}{Base for tree prior on tau(x) trees (see details)}

\item{power_moderate}{Power for the tree prior on tau(x) trees (see details)}

\item{nu}{Degrees of freedom in the chisq prior on \eqn{sigma^2}}

\item{lambda}{Scale parameter in the chisq prior on \eqn{sigma^2}}

\item{sigq}{Calibration quantile for the chisq prior on \eqn{sigma^2}}

\item{sighat}{Calibration estimate for the chisq prior on \eqn{sigma^2}}

\item{include_pi}{Takes values "control", "moderate", "both" or "none". Whether to
include pihat in mu(x) ("control"), tau(x) ("moderate"), both or none. Values of "control"
or "both" are HIGHLY recommended with observational data.}

\item{use_muscale}{Use a half-Cauchy hyperprior on the scale of mu.}

\item{use_tauscale}{Use a half-Normal prior on the scale of tau.}

\item{n_cores}{An optional integer of the number of cores to run your warm start-BCF.}

\item{warm_start_fit}{An XBCF fit object, that will be used to initialize warm start-BCF.}

\item{cutpoint_grid}{The scheme for cutpoint grid generation:
"unifrom" [default] option generates a unifromly distributed cutpoint grid for each continuous variable (size is 10*number of observations);
"exact" option uses sets of unique values at each categorical or continuous variable as availible cutpoints.}
}
\value{
A list with elements
\item{tau}{\code{nsim} by \code{n} matrix of posterior samples of individual treatment effects}
\item{yhat}{\code{nsim} by \code{n} matrix of posterior samples of individual outcome}
}
\description{
Fit warm-start Bayesian Causal Forests
}
\details{
Fits multiple Bayesian Causal Forest models (Hahn et. al., 2020),
that are independently initialized with forests extracted from a fitted Accelerated Bayesian Causal Forests heuristic (Krantsevich et al., 2023).
For a response variable y, binary treatment z, and covariates x,
\deqn{y_i = a \mu(x_i, \pi_i) + b_{z_i}\tau(x_i, \pi_i) + \epsilon_i}
where \eqn{\pi_i} is an estimate of the propensity score \eqn{\Pr(Z_i=1 | X_i=x_i)};
\eqn{a, b_0, b_1} are scaling parameters; and \eqn{\epsilon_i \sim N(0,\sigma^2)}.

Some notes:
\itemize{
   \item x_control and x_moderate must be numeric matrices. See e.g. the makeModelMatrix function in the
   dbarts package for appropriately constructing a design matrix from a data.frame.
   For proper initialization, these matrices must be identical to the ones used when fitting XBCF model.
   \item sd_control and sd_moderate are the prior SD(mu(x)) and SD(tau(x)) at a given value of x (respectively). If
   use_muscale = FALSE, then this is the parameter \eqn{\sigma_\mu} from the original BART paper, where the leaf parameters
   have prior distribution \eqn{N(0, \sigma_\mu/m)}, where m is the number of trees.
   If use_muscale=TRUE then sd_control is the prior median of a half Cauchy prior for SD(mu(x)). If use_tauscale = TRUE,
   then sd_moderate is the prior median of a half Normal prior for SD(tau(x)).
   \item By default the prior on \eqn{\sigma^2} is calibrated as in Chipman, George and McCulloch (2008).

}
}
\references{
Hahn, Murray, and Carvalho (2020). Bayesian Regression Tree Models for Causal Inference: Regularization, Confounding, and Heterogeneous Effects (with Discussion).
Bayesian Anal. 15(3): 965-1056 (September 2020). DOI: 10.1214/19-BA1195
}
